+++ src/CommandLine/OptionValueProviders.cs
@@ -297,6 +297,15 @@ internal static class OptionValueProviders
         OptionValues.TrimStart
     );
 
+    public static OptionValueProvider FunctionFlagsProvider { get; } = new(
+        MetaValues.Function,
+        OptionValues.FunctionFlags_Sort,
+        OptionValues.FunctionFlags_SortDescending,
+        SimpleOptionValue.Create(ModifyFlags.Distinct, shortValue: "", description: "Return distinct values."),
+        OptionValues.ModifyFlags_Group,
+        OptionValues.SortBy
+    );
+
     public static OptionValueProvider HighlightOptionsProvider { get; } = new(
         MetaValues.Highlight,
         OptionValues.HighlightOptions_None,