+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add option `--function` which replaces `--modify` ([#87](https://github.com/josefpihrt/orang/pull/87)).
+
+### Changed
+
+- Make option `--modify` obsolete ([#87](https://github.com/josefpihrt/orang/pull/87)).
+
 ### Fixed
 
 - [CLI] Fix reading of redirected input ([PR](https://github.com/dotnet/roslynator/pull/91))